requirements.txt
python-dotenv==1.0.1
streamlit==1.36.0

langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4

langchain-groq==0.1.6
langchain-huggingface==0.0.3

pinecone-client==3.0.0
wikipedia
arxiv
pypdf

torch
sentence-transformers